+++ MacroPepelelipa/MacroPepelelipa/Model/Colors.swift
@@ -1,0 +1,19 @@
+//
+//  Colors.swift
+//  MacroPepelelipa
+//
+//  Created by Pedro Giuliano Farina on 30/09/20.
+//  Copyright © 2020 Pedro Giuliano Farina. All rights reserved.
+//
+
+import UIKit
+
+public extension UIColor {
+    static let actionColor = UIColor(named: "Action")
+    static let backgroundColor = UIColor(named: "Background")
+    static let bodyColor = UIColor(named: "Body")
+    static let placeholderColor = UIColor(named: "Placeholder")
+    static let rootColor = UIColor(named: "Root")
+    static let titleColor = UIColor(named: "Title")
+    static let toolsColor = UIColor(named: "Tools")
+}
