+++ MacroPepelelipa/MacroPepelelipa/View/Colors.xcassets/Background.colorset/Contents.json
